+++ telecom/gr-limesdr/lib/source_impl.h
@@ -102,7 +102,7 @@ public:
 
     void write_lms_reg(uint32_t address, uint16_t val);
 
-    void set_dspcfg_preamble(uint16_t dspcfg_PASSTHROUGH_LEN = 100u, uint16_t dspcfg_THRESHOLD = 100u, int dspcfg_preamble_en = 0);
+    void set_dspcfg_preamble(uint16_t dspcfg_PASSTHROUGH_LEN = 100u, uint8_t dspcfg_THRESHOLD = 100u, int dspcfg_preamble_en = 0);
 
     uint32_t get_dspcfg_short_sum() ;
